pyproject.toml
[tool.poetry]
name = "snowflake-test"
version = "0.1.0"
description = ""
authors = ["Igor Serko <igor@ly.st>"]

[tool.poetry.dependencies]
python = "^3.7"
snowflake-connector-python = { version = "^2.1", extras = ["pandas"] }
environs = "^7.1.0"

[tool.poetry.dev-dependencies]
black = { version = "*", allow-prereleases = true }
flake8 = "^3.7.9"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.black]
line-length = 79
target-version = ["py37"]
